feat: implement attendance system core
- add students, sessions, events schema (migrations)
- implement repositories for attendance logic
- add service layer for business logic
- update HTTP handlers and DTOs
- wire dependencies in main.go

+++ internal/transport/http/dto/attendance_responce.go
@@ -1,7 +1,7 @@
 package dto
 
 type AttendanceResponse struct {
-	SessionID    int      `json:"session_id"`
+	SessionID    int64    `json:"session_id"`
 	SavedEvents  int      `json:"saved_events"`
 	NotFoundCard []string `json:"not_found_card"`
 }
